backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pillow==10.0.0
segment-anything==1.0
python-dotenv==1.0.0
opencv-python==4.8.0.76
aiofiles==23.1.0 